+++ lib/widgets/adaptive_dialog.dart
@@ -2,6 +2,8 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:pro_image_editor/utils/design_mode.dart';
 
+import '../models/theme/theme.dart';
+
 /// A dialog that adapts its appearance based on the design mode.
 class AdaptiveDialog extends StatefulWidget {
   /// The design mode to determine the appearance of the dialog.
@@ -19,12 +21,15 @@ class AdaptiveDialog extends StatefulWidget {
   /// The brightness of the dialog.
   final Brightness brightness;
 
+  final ImageEditorTheme imageEditorTheme;
+
   /// Creates an [AdaptiveDialog].
   ///
   /// The [designMode] determines the appearance of the dialog.
   /// The [title] and [content] are required widgets to display in the dialog.
   /// The [actions] is a list of [AdaptiveDialogAction] widgets to include as buttons.
   /// The [brightness] controls the brightness of the dialog.
+  /// The [imageEditorTheme] is the theme specific to the image editor.
   const AdaptiveDialog({
     super.key,
     required this.designMode,
@@ -32,6 +37,7 @@ class AdaptiveDialog extends StatefulWidget {
     required this.content,
     required this.actions,
     required this.brightness,
+    required this.imageEditorTheme,
   });
 
   @override
@@ -45,8 +51,23 @@ class _AdaptiveDialogState extends State<AdaptiveDialog> {
     if (widget.designMode == ImageEditorDesignModeE.cupertino) {
       // Return a Cupertino-style dialog when in Cupertino design mode.
       return CupertinoTheme(
-        data:
-            CupertinoTheme.of(context).copyWith(brightness: widget.brightness),
+        data: CupertinoTheme.of(context).copyWith(
+          brightness: widget.brightness,
+          primaryColor: widget.brightness == Brightness.dark
+              ? widget.imageEditorTheme.adaptiveDialogTheme
+                  .cupertinoPrimaryColorDark
+              : widget.imageEditorTheme.adaptiveDialogTheme
+                  .cupertinoPrimaryColorLight,
+          textTheme: CupertinoTextThemeData(
+            textStyle: TextStyle(
+              color: widget.brightness == Brightness.dark
+                  ? widget.imageEditorTheme.adaptiveDialogTheme
+                      .cupertinoPrimaryColorDark
+                  : widget.imageEditorTheme.adaptiveDialogTheme
+                      .cupertinoPrimaryColorLight,
+            ),
+          ),
+        ),
         child: CupertinoAlertDialog(
           title: widget.title,
           content: widget.content,